pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssis-adf-agent"
version = "0.1.0"
description = "MCP server for reading SSIS packages and converting them to ADF pipelines, surfaced as GitHub Copilot tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "azure-mgmt-datafactory>=4.0.0",
    "azure-mgmt-web>=7.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-applicationinsights>=4.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-identity>=1.15.0",
    "pyodbc>=5.0.0",
    "gitpython>=3.1.40",
    "jsonschema>=4.20.0",
    "networkx>=3.2.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "lxml-stubs>=0.5.0",
]
llm = [
    "openai>=1.30.0",
]
pdf = [
    "reportlab>=4.0.0",
]

[project.scripts]
ssis-adf-agent = "ssis_adf_agent.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["ssis_adf_agent"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
# UP042: don't auto-migrate `class X(str, Enum)` to `StrEnum` — changes
# `str(X.MEMBER)` from "X.MEMBER" to "MEMBER" and risks breaking callers.
ignore = ["UP042"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
